requirements.txt
flask~=1.0.3
flask_restful
flask_cors
flask_sqlalchemy
flask_jwt_extended
flask_uploads
marshmallow==3.0.0rc5
flask_marshmallow
marshmallow-sqlalchemy
minio
Pillow
requests
gunicorn
python-dotenv~=0.10.3

SQLAlchemy~=1.3.0
Werkzeug~=0.15.4